tests(bot): add tests for `roll` command
Added tests for `rol` command using `dpytest`.

Changes to be committed:
	modified:   bot/cogs/__init__.py
	modified:   bot/cogs/misc.py
	new file:   bot/cogs/rol.py
	modified:   tests/bot/cogs/test_misc.py
	modified:   tests/chat/test_models.py
