+++ bot/cogs/__init__.py
@@ -1,4 +1,5 @@
-from .misc import Miscellaneous, Roleplay
+from .misc import Miscellaneous
+from .rol import Roleplay
 
 __all__ = [
     'Miscellaneous', 'Roleplay'
